refactor: copy paste plugin
