+++ frontend/appflowy_flutter/lib/plugins/document/presentation/editor_plugins/copy_and_paste/copy_and_paste_plugin.dart
@@ -1,0 +1,2 @@
+export 'commands/copy_and_paste_commands.dart';
+export 'clipboard_service.dart';
