+++ frontend/appflowy_flutter/lib/plugins/document/presentation/editor_plugins/copy_and_paste/editor_state_extensions/editor_state_extensions.dart
@@ -1,0 +1,5 @@
+export 'paste_from_html.dart';
+export 'paste_from_image.dart';
+export 'paste_from_in_app_json.dart';
+export 'paste_from_plain_text.dart';
+export 'paste_node_extension.dart';
