+++ frontend/appflowy_flutter/lib/plugins/document/presentation/editor_plugins/copy_and_paste/editor_state_extensions/paste_from_html.dart
@@ -1,4 +1,4 @@
-import 'package:appflowy/plugins/document/presentation/editor_plugins/copy_and_paste/editor_state_paste_node_extension.dart';
+import 'package:appflowy/plugins/document/presentation/editor_plugins/copy_and_paste/editor_state_extensions/paste_node_extension.dart';
 import 'package:appflowy_editor/appflowy_editor.dart';
 
 extension PasteFromHtml on EditorState {
