+++ frontend/appflowy_flutter/lib/plugins/document/presentation/editor_plugins/plugins.dart
@@ -6,9 +6,7 @@ export 'callout/callout_block_component.dart';
 export 'code_block/code_block_component.dart';
 export 'code_block/code_block_shortcut_event.dart';
 export 'context_menu/custom_context_menu.dart';
-export 'copy_and_paste/custom_copy_command.dart';
-export 'copy_and_paste/custom_cut_command.dart';
-export 'copy_and_paste/custom_paste_command.dart';
+export 'copy_and_paste/commands/copy_and_paste_commands.dart';
 export 'database/database_view_block_component.dart';
 export 'database/inline_database_menu_item.dart';
 export 'database/referenced_database_menu_item.dart';
