+++ frontend/appflowy_flutter/test/bloc_test/shortcuts_test/shortcuts_cubit_test.dart
@@ -73,7 +73,7 @@ void main() {
               .having(
                 (w) => w.commandShortcutEvents,
                 'shortcuts',
-                commandShortcutEvents,
+                customizableShortcutEvents,
               ),
         ],
       );
@@ -155,7 +155,7 @@ void main() {
               .having(
                 (w) => w.commandShortcutEvents,
                 'shortcuts',
-                commandShortcutEvents,
+                customizableShortcutEvents,
               ),
         ],
       );
